ludamind/requirements.txt
# ╔══════════════════════════════════════════════════════════════╗
# ║          TrendsPro - Clean Architecture Requirements          ║
# ║                     FastAPI + DDD + SOLID                     ║
# ╚══════════════════════════════════════════════════════════════╝

# =================================================================
# Core Web Framework
# =================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# =================================================================
# API & Validation
# =================================================================
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# =================================================================
# Database Drivers
# =================================================================
# MySQL
mysql-connector-python==8.2.0
aiomysql==0.2.0

# MongoDB
pymongo==4.6.1
motor==3.3.2  # Async MongoDB driver

# Redis (for caching)
redis==5.0.1
aioredis==2.0.1

# =================================================================
# AI/LLM Integration
# =================================================================
openai==1.9.0
tiktoken==0.5.2  # Token counting for OpenAI
langchain==0.1.0  # Optional: for advanced LLM chains

# =================================================================
# Authentication & Security
# =================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# =================================================================
# HTTP & Networking
# =================================================================
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0
urllib3==2.1.0

# =================================================================
# Data Processing
# =================================================================
pandas==2.1.4
numpy==1.26.3

# =================================================================
# Utilities
# =================================================================
python-dateutil==2.8.2
pytz==2023.3
typing-extensions==4.9.0

# =================================================================
# Logging & Monitoring
# =================================================================
structlog==24.1.0
colorama==0.4.6
rich==13.7.0  # Beautiful terminal output

# =================================================================
# Testing
# =================================================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx-mock==0.5.0

# =================================================================
# Development Tools (Optional)
# =================================================================
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pre-commit==3.6.0

# =================================================================
# Documentation
# =================================================================
mkdocs==1.5.3
mkdocs-material==9.5.3

# =================================================================
# Metrics & Observability
# =================================================================
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# =================================================================
# Task Queue (Optional for background jobs)
# =================================================================
celery==5.3.4
flower==2.0.1  # Celery monitoring

# =================================================================
# Migration from Flask (can be removed after full migration)
# =================================================================
flask==3.0.0
werkzeug==3.0.1
flask-cors==4.0.0
PyJWT==2.8.0
